Update GitHub Actions workflow for Azure deployment
This commit modifies the GitHub Actions workflow to streamline the deployment process. Changes include updating the names and settings of individual steps, as well as altering the process of building and pushing the Docker image to Azure. Additionally, secrets for values such as registry login server and username have been updated for consistency.